get rid of beanstalkutils folder
