--- Tasks/Common/beanstalkutils/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "beanstalkutils",
-  "version": "1.0.0",
-  "description": "Shared routines for working with Elastic Beanstalk deployments",
-  "main": "beanstalkutils.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Amazon Web Services",
-  "license": "MIT"
-}
