--- Tasks/Common/beanstalkutils/tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES6",
-        "module": "commonjs",
-        "sourceMap": true,
-        "alwaysStrict": true,
-        "declaration": true
-    },
-    "exclude": [
-        "node_modules"
-    ]
-}
